update perf param and delete minimax testcases
