+++ .github/workflows/full-test-npu.yml
@@ -673,11 +673,11 @@ jobs:
             test_case: test/registered/ascend/performance/deepseek_v3_2/test_npu_deepseek_v3_2_w8a8_1p1d_32p_in128k_out1k_bs8.py
             test_type: 'perf'
             prefill_decode_deployment: 'separation'
-          - name: glm5_1_w4a8_1p1d_32p_in128k_out1k
+          - name: glm5_1_w4a8_1p1d_32p_in128k_out1k_20ms
             prefill_size: 2
             decode_size: 2
             router_size: 1
-            test_case: test/registered/ascend/performance/glm5_1/test_npu_glm5_1_w4a8_1p1d_32p_in128k_out1k.py
+            test_case: test/registered/ascend/performance/glm5_1/test_npu_glm5_1_w4a8_1p1d_32p_in128k_out1k_20ms.py
             test_type: 'perf'
             prefill_decode_deployment: 'separation'
           - name: glm5_1_w4a8_1p1d_32p_in16k_out1k_50ms
