+++ .github/workflows/nightly-test-npu.yml
@@ -223,13 +223,13 @@ jobs:
             test_case: test/registered/ascend/performance/qwen3_next_80b_a3b_instruct/test_npu_qwen3_next_80b_w8a8_2p_in6k_out1k5_bs16_aime25.py
             test_type: 'perf'
           # minimax_m2_5 performance tests
-          - name: minimax_m2_5_w8a8_8p_in3k5_out1k5_50ms_aime25
+          - name: minimax_m2_5_w8a8_8p_in3k5_out1k5_50ms_gpqa
             runner: linux-aarch64-a3-16
-            test_case: test/registered/ascend/performance/minimax_m2_5/test_npu_minimax_m2_5_w8a8_8p_in3k5_out1k5_50ms_aime25.py
+            test_case: test/registered/ascend/performance/minimax_m2_5/test_npu_minimax_m2_5_w8a8_8p_in3k5_out1k5_50ms_gpqa.py
             test_type: 'perf'
-          - name: minimax_m2_5_w8a8_4p_in64k_out1k_prefix90_50ms_aime25
+          - name: minimax_m2_5_w8a8_4p_in64k_out1k_prefix90_50ms_gpqa
             runner: linux-aarch64-a3-16
-            test_case: test/registered/ascend/performance/minimax_m2_5/test_npu_minimax_m2_5_w8a8_4p_in64k_out1k_prefix90_50ms_aime25.py
+            test_case: test/registered/ascend/performance/minimax_m2_5/test_npu_minimax_m2_5_w8a8_4p_in64k_out1k_prefix90_50ms_gpqa.py
             test_type: 'perf'
           # deepseek_v3_2 accuracy tests
           - name: deepseek_v3_2_8p_aime25
