+++ test/registered/ascend/performance/minimax_m2_5/test_npu_minimax_m2_5_w8a8_8p_in128k_out1k_prefix90_20ms.py
@@ -100,7 +100,7 @@ class TestNPUMiniMaxM2_5W8A8_8P_In128k_Out1k_Prefix90_20ms(
     output_len = 1024
     random_range_ratio = 1
     repeat_rate = 0.9
-    tpot = 20
+    tpot = 24.44
     output_token_throughput = 66.29
     request_rate = float("inf")
