+++ test/registered/ascend/performance/qwen3_5_397b/test_npu_qwen3_5_397b_w4a8_8p_in64k_out1k_prefix90_50ms.py
@@ -19,7 +19,8 @@ QWEN3_5_397B_64K_PREFIX_ENVS = {
     "SGLANG_SET_CPU_AFFINITY": "1",
     "ASCEND_USE_FIA": "1",
     "SGLANG_DEEPEP_NUM_MAX_DISPATCH_TOKENS_PER_RANK": "128",
-    "HCCL_BUFFSIZE": "0",
+    "HCCL_BUFFSIZE": "64",
+    "ZBAL_HCCL_OP": "allreduce,_allgather_base,allgather,broadcast,scatter,reduce_scatter,_reduce_scatter_base,alltoall_base",
     "DEEPEP_NORMAL_LONG_SEQ_ROUND": "20",
     "DEEPEP_NORMAL_LONG_SEQ_PER_ROUND_TOKENS": "4096",
     "DEEP_NORMAL_MODE_USE_INT8_QUANT": "1",
@@ -113,8 +114,8 @@ class TestNPUQwen3_5_397B_64K_Prefix90(TestAscendPerformanceTestCaseBase):
     envs = QWEN3_5_397B_64K_PREFIX_ENVS
     dataset_name = "generated-shared-prefix"
     warmup_requests = 0
-    max_concurrency = 128
-    num_prompts = 128
+    max_concurrency = 112
+    num_prompts = 112
     repeat_rate = 0.9
     input_len = 65536
     output_len = 1024
