+++ test/registered/ascend/performance/qwen3_6_35b_a3b/test_npu_qwen3_6_35b_a3b_1p_in64k_out1k_prefix90_50ms_aime26.py
@@ -9,14 +9,6 @@ from sglang.test.ascend.e2e.test_npu_performance_utils import (
     QWEN3_6_35B_A3B_MODEL_PATH,
     TestAscendPerformanceTestCaseBase,
 )
-from sglang.test.ci.ci_register import register_npu_ci
-
-register_npu_ci(
-    est_time=3600,
-    suite="",
-    nightly=True,
-    disabled="performance testcase",
-)
 
 QWEN3_6_35B_A3B_64K_PREFIX_ENVS = {
     "PYTORCH_NPU_ALLOC_CONF": "expandable_segments:True",
@@ -46,11 +38,10 @@ QWEN3_6_35B_A3B_64K_PREFIX_OTHER_ARGS = [
     "--max-prefill-tokens",
     65536,
     "--trust-remote-code",
-    "--enable-prefill-delayer",
     "--mamba-scheduler-strategy",
     "extra_buffer",
     "--max-running-requests",
-    42,
+    40,
     "--max-mamba-cache-size",
     210,
     "--mem-fraction-static",
@@ -63,7 +54,6 @@ QWEN3_6_35B_A3B_64K_PREFIX_OTHER_ARGS = [
     32,
     36,
     40,
-    42,
     "--enable-multimodal",
     "--mm-attention-backend",
     "ascend_attn",
@@ -118,8 +108,8 @@ class TestNPUQwen3_6_35BA3B_1P_In64k_Out1k_Prefix90_50ms(
     other_args = QWEN3_6_35B_A3B_64K_PREFIX_OTHER_ARGS
     envs = QWEN3_6_35B_A3B_64K_PREFIX_ENVS
     dataset_name = "generated-shared-prefix"
-    max_concurrency = 42
-    num_prompts = 42
+    max_concurrency = 40
+    num_prompts = 40
     input_len = 65536
     output_len = 1024
     random_range_ratio = 1
